CLN: 컨트롤러에 합쳐져 있는 비즈니스 로직 분리
- SOLID 원칙의 S 단일 책임의 원칙에 따라 한 개의 클래스는 단 하나의 책임만 지도록 분리하였습니다.
