+++ druginfo/src/main/java/com/dcm/spring/druginfo/controller/DrugOpenApiController.java
@@ -1,6 +1,8 @@
 package com.dcm.spring.druginfo.controller;
 
 import com.dcm.spring.druginfo.model.DrugResponseDto;
+import com.dcm.spring.druginfo.service.DrugService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -14,44 +16,15 @@ import java.nio.charset.StandardCharsets;
 
 @RestController
 public class DrugOpenApiController {
-    private final static String BASE_URL = "http://apis.data.go.kr/1471000/DrbEasyDrugInfoService/getDrbEasyDrugList";
-    private final String API_KEY = "2SKRXQ8IZz8FeCOyzzs8hgyk0YWybv0fp5wAMHnsnrgPTiEqwl%2FxxQSs%2BZMHxjUSUxZWIQ37pEh9zUjkqh9zlg%3D%3D";
+    private final DrugService drugService;
 
-    private Mono<Object> convertToMedication(DrugResponseDto drugResponseDTO) {
-        return Mono.just(drugResponseDTO.getBody().getItems().get(0));
+    @Autowired
+    public DrugOpenApiController(DrugService drugService) {
+        this.drugService = drugService;
     }
+
     @GetMapping(value = "/PublicData/{itemName}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public Mono<Object> getInfo(@PathVariable String itemName) {
-        DefaultUriBuilderFactory factory = new DefaultUriBuilderFactory(BASE_URL);
-        factory.setEncodingMode(DefaultUriBuilderFactory.EncodingMode.VALUES_ONLY);
-        String encodedItemName = URLEncoder.encode(itemName, StandardCharsets.UTF_8);
-        WebClient webClient = WebClient.builder()
-                .uriBuilderFactory(factory)
-                .baseUrl(BASE_URL)
-                .build();
-
-        Mono<DrugResponseDto> response = webClient.get()
-                .uri(uriBuilder -> uriBuilder
-                        .queryParam("serviceKey", API_KEY)
-                        .queryParam("pageNo", "1")
-                        .queryParam("numOfRows", "1")
-                        .queryParam("entpName", "")
-                        .queryParam("itemName", encodedItemName)
-                        .queryParam("itemSeq", "")
-                        .queryParam("efcyQesitm", "")
-                        .queryParam("useMethodQesitm", "")
-                        .queryParam("atpnWarnQesitm", "")
-                        .queryParam("atpnQesitm", "")
-                        .queryParam("intrcQesitm", "")
-                        .queryParam("seQesitm", "")
-                        .queryParam("depositMethodQesitm", "")
-                        .queryParam("openDe", "")
-                        .queryParam("updateDe", "")
-                        .queryParam("type", "json")
-                        .build())
-                .retrieve()
-                .bodyToMono(DrugResponseDto.class);
-
-        return response.flatMap(this::convertToMedication);
+    public Mono<Object> getDrugInfo(@PathVariable String itemName){
+        return drugService.getInfo(itemName);
     }
 }
